+++ addons/pvr.mythtv.cmyth/src/cppmyth/MythRecorder.h
@@ -44,7 +44,7 @@ public:
   void Lock();
   void Unlock();
 
-  int ID();
+  unsigned int ID();
   MythProgramInfo GetCurrentProgram();
 
   bool IsRecording();
@@ -56,7 +56,7 @@ public:
   bool LiveTVWatch(const CStdString &msg);
   bool LiveTVDoneRecording(const CStdString &msg);
   bool LiveTVChainUpdate(const CStdString &chainid);
-  int ReadLiveTV(void *buffer, unsigned long length);
+  int ReadLiveTV(void *buffer, unsigned int length);
   long long LiveTVSeek(long long offset, int whence);
   long long LiveTVDuration();
